src/consumers/consumer_requirements.txt
pydantic
minio
python-dotenv
confluent-kafka
tomli
pillow
torch
torchvision